+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 3.3.0-RC1
 + PLGOPNS-247: [refactor] Extend compatibility of the extension for OpenCart v. 2.X
 + PLGOPNS-315: [refactor] Fix typo error of warning message in settings page
++ PLGOPNS-303: [refactor] Improve iDEAL issuer selection list
 
 ## 3.2.0
 Release date: September, 17th 2020